bq27441: Factory reset option, polarity clean-up
Include the lowBat polarity in the debugfs list

Add a file to be able to force factory reset. This
mode will set the bq27441 to run from OTP flash,
which means default settings for all parameters.
This is for testing in the lab, to compare new and
old configuration.

The DMCode will be 255 for factory reset.